refactor - #122 controller가 하나의 class를 바라보도록 변경
