+++ src/main/java/sopt/org/hmh/domain/dailychallenge/controller/DailyChallengeController.java
@@ -8,7 +8,6 @@ import sopt.org.hmh.domain.dailychallenge.domain.exception.DailyChallengeSuccess
 import sopt.org.hmh.domain.dailychallenge.dto.request.FinishedDailyChallengeListRequest;
 import sopt.org.hmh.domain.dailychallenge.dto.request.FinishedDailyChallengeStatusListRequest;
 import sopt.org.hmh.domain.dailychallenge.service.DailyChallengeFacade;
-import sopt.org.hmh.domain.dailychallenge.service.DailyChallengeService;
 import sopt.org.hmh.global.auth.UserId;
 import sopt.org.hmh.global.common.response.BaseResponse;
 import sopt.org.hmh.global.common.response.EmptyJsonResponse;
@@ -19,7 +18,6 @@ import sopt.org.hmh.global.common.response.EmptyJsonResponse;
 public class DailyChallengeController implements DailyChallengeApi {
 
     private final DailyChallengeFacade dailyChallengeFacade;
-    private final DailyChallengeService dailyChallengeService;
 
     @Override
     @PostMapping("/finish")
@@ -41,7 +39,7 @@ public class DailyChallengeController implements DailyChallengeApi {
             @RequestHeader("OS") final String os,
             @RequestBody final FinishedDailyChallengeStatusListRequest request
     ) {
-        dailyChallengeService.changeDailyChallengeStatusByIsSuccess(userId, request);
+        dailyChallengeFacade.changeDailyChallengeStatusByIsSuccess(userId, request);
         return ResponseEntity
                 .status(DailyChallengeSuccess.SEND_FINISHED_DAILY_CHALLENGE_SUCCESS.getHttpStatus())
                 .body(BaseResponse.success(DailyChallengeSuccess.SEND_FINISHED_DAILY_CHALLENGE_SUCCESS, new EmptyJsonResponse()));
